Modify StateCacher, add checking to DescriptorSetWithOffsets constructor
